+++ src/main/java/com/listeny/listeny/repository/SeguidoresRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SeguidoresRepository extends JpaRepository<Seguidores, Integer> {
+public interface SeguidoresRepository extends JpaRepository<Seguidores, Long> {
 
 }
